Checkpoint work

+++ internal/log_analysis/log_processor/common/common.go
@@ -52,7 +52,7 @@ type EnvConfig struct {
 	ProcessedDataBucket         string `required:"true" split_words:"true"`
 	SqsQueueURL                 string `required:"true" split_words:"true"`
 	SnsTopicARN                 string `required:"true" split_words:"true"`
-	TimeLimitSec                int    `required:"true" split_words:"true"`
+	TimeLimitSec                int64  `required:"true" split_words:"true"`
 }
 
 func SessionInit() {
